// frontend/src/api/generated/models/sprintDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * acpms-server
 * OpenAPI spec version: 0.1.0
 */

export interface SprintDto {
  /** @nullable */
  closed_at?: string | null;
  /** @nullable */
  closed_by?: string | null;
  created_at: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  end_date?: string | null;
  /** @nullable */
  goal?: string | null;
  id: string;
  name: string;
  project_id: string;
  sequence: number;
  /** @nullable */
  start_date?: string | null;
  status: string;
  updated_at: string;
}